+++ pkg/proxy/proxy.go
@@ -20,6 +20,7 @@ import (
 	"bytes"
 	"fmt"
 	"net/http"
+	"net/textproto"
 	"strings"
 	"text/template"
 
@@ -126,7 +127,7 @@ type krpAuthorizerAttributesGetter struct {
 
 // GetRequestAttributes populates authorizer attributes for the requests to kube-rbac-proxy.
 func (n krpAuthorizerAttributesGetter) GetRequestAttributes(u user.Info, r *http.Request) []authorizer.Attributes {
-	apiVerb := ""
+	apiVerb := "*"
 	switch r.Method {
 	case "POST":
 		apiVerb = "create"
@@ -187,8 +188,10 @@ func (n krpAuthorizerAttributesGetter) GetRequestAttributes(u user.Info, r *http
 		}
 	}
 	if n.authzConfig.Rewrites.ByHTTPHeader != nil && n.authzConfig.Rewrites.ByHTTPHeader.Name != "" {
-		if p := r.Header.Get(n.authzConfig.Rewrites.ByHTTPHeader.Name); p != "" {
-			params = append(params, p)
+		mimeHeader := textproto.MIMEHeader(r.Header)
+		mimeKey := textproto.CanonicalMIMEHeaderKey(n.authzConfig.Rewrites.ByHTTPHeader.Name)
+		if ps, ok := mimeHeader[mimeKey]; ok {
+			params = append(params, ps...)
 		}
 	}
 
@@ -211,49 +214,6 @@ func (n krpAuthorizerAttributesGetter) GetRequestAttributes(u user.Info, r *http
 		allAttrs = append(allAttrs, attrs)
 	}
 	return allAttrs
-}
-
-// DeepCopy of Proxy Configuration
-func (c *Config) DeepCopy() *Config {
-	res := &Config{
-		Authentication: &authn.AuthnConfig{},
-	}
-
-	if c.Authentication != nil {
-		res.Authentication = &authn.AuthnConfig{}
-
-		if c.Authentication.X509 != nil {
-			res.Authentication.X509 = &authn.X509Config{
-				ClientCAFile: c.Authentication.X509.ClientCAFile,
-			}
-		}
-
-		if c.Authentication.Header != nil {
-			res.Authentication.Header = &authn.AuthnHeaderConfig{
-				Enabled:         c.Authentication.Header.Enabled,
-				UserFieldName:   c.Authentication.Header.UserFieldName,
-				GroupsFieldName: c.Authentication.Header.GroupsFieldName,
-				GroupSeparator:  c.Authentication.Header.GroupSeparator,
-			}
-		}
-	}
-
-	if c.Authorization != nil {
-		if c.Authorization.ResourceAttributes != nil {
-			res.Authorization = &authz.Config{
-				ResourceAttributes: &authz.ResourceAttributes{
-					Namespace:   c.Authorization.ResourceAttributes.Namespace,
-					APIGroup:    c.Authorization.ResourceAttributes.APIGroup,
-					APIVersion:  c.Authorization.ResourceAttributes.APIVersion,
-					Resource:    c.Authorization.ResourceAttributes.Resource,
-					Subresource: c.Authorization.ResourceAttributes.Subresource,
-					Name:        c.Authorization.ResourceAttributes.Name,
-				},
-			}
-		}
-	}
-
-	return res
 }
 
 func templateWithValue(templateString, value string) string {