+++ test/kubetest/kubernetes.go
@@ -30,7 +30,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -38,7 +37,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 )
 
-func CreatedManifests(client kubernetes.Interface, paths ...string) Setup {
+func CreatedManifests(client kubernetes.Interface, paths ...string) Action {
 	return func(ctx *ScenarioContext) error {
 		for _, path := range paths {
 			content, err := ioutil.ReadFile(path)
@@ -105,7 +104,7 @@ func createClusterRole(client kubernetes.Interface, ctx *ScenarioContext, conten
 
 	_, err := client.RbacV1().ClusterRoles().Create(context.TODO(), cr, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.RbacV1().ClusterRoles().Delete(context.TODO(), cr.Name, metav1.DeleteOptions{})
 	})
 
@@ -122,7 +121,7 @@ func createClusterRoleBinding(client kubernetes.Interface, ctx *ScenarioContext,
 
 	_, err := client.RbacV1().ClusterRoleBindings().Create(context.TODO(), crb, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.RbacV1().ClusterRoleBindings().Delete(context.TODO(), crb.Name, metav1.DeleteOptions{})
 	})
 
@@ -141,7 +140,7 @@ func createDeployment(client kubernetes.Interface, ctx *ScenarioContext, content
 
 	_, err := client.AppsV1().Deployments(d.Namespace).Create(context.TODO(), &d, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		dep, err := client.AppsV1().Deployments(d.Namespace).Get(context.TODO(), d.Name, metav1.GetOptions{})
 		if err != nil {
 			return err
@@ -185,7 +184,7 @@ func dumpLogs(client kubernetes.Interface, ctx *ScenarioContext, opts metav1.Lis
 				return
 			}
 
-			io.Copy(os.Stdout, stream)
+			_, _ = io.Copy(os.Stdout, stream)
 		}
 	}
 }
@@ -202,7 +201,7 @@ func createService(client kubernetes.Interface, ctx *ScenarioContext, content []
 
 	_, err := client.CoreV1().Services(s.Namespace).Create(context.TODO(), s, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.CoreV1().Services(s.Namespace).Delete(context.TODO(), s.Name, metav1.DeleteOptions{})
 	})
 
@@ -221,7 +220,7 @@ func createServiceAccount(client kubernetes.Interface, ctx *ScenarioContext, con
 
 	_, err := client.CoreV1().ServiceAccounts(sa.Namespace).Create(context.TODO(), sa, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.CoreV1().ServiceAccounts(sa.Namespace).Delete(context.TODO(), sa.Name, metav1.DeleteOptions{})
 	})
 
@@ -240,7 +239,7 @@ func createSecret(client kubernetes.Interface, ctx *ScenarioContext, content []b
 
 	_, err := client.CoreV1().Secrets(secret.Namespace).Create(context.TODO(), secret, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.CoreV1().Secrets(secret.Namespace).Delete(context.TODO(), secret.Name, metav1.DeleteOptions{})
 	})
 
@@ -259,7 +258,7 @@ func createConfigmap(client kubernetes.Interface, ctx *ScenarioContext, content 
 
 	_, err := client.CoreV1().ConfigMaps(configmap.Namespace).Create(context.TODO(), configmap, metav1.CreateOptions{})
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		return client.CoreV1().ConfigMaps(configmap.Namespace).Delete(context.TODO(), configmap.Name, metav1.DeleteOptions{})
 	})
 
@@ -362,26 +361,19 @@ func podRunningAndReady(pod corev1.Pod) (bool, error) {
 	return false, nil
 }
 
-func Sleep(d time.Duration) Condition {
-	return func(ctx *ScenarioContext) error {
-		time.Sleep(d)
-		return nil
-	}
-}
-
 type RunOptions struct {
 	ServiceAccount     string
 	TokenAudience      string
 	ClientCertificates bool
 }
 
-func RunSucceeds(client kubernetes.Interface, image string, name string, command []string, opts *RunOptions) Check {
+func RunSucceeds(client kubernetes.Interface, image string, name string, command []string, opts *RunOptions) Action {
 	return func(ctx *ScenarioContext) error {
 		return run(client, ctx, image, name, command, opts)
 	}
 }
 
-func RunFails(client kubernetes.Interface, image string, name string, command []string, opts *RunOptions) Check {
+func RunFails(client kubernetes.Interface, image string, name string, command []string, opts *RunOptions) Action {
 	return func(ctx *ScenarioContext) error {
 		err := run(client, ctx, image, name, command, opts)
 		if err == nil {
@@ -482,7 +474,7 @@ func run(client kubernetes.Interface, ctx *ScenarioContext, image string, name s
 		return fmt.Errorf("failed to create job: %v", err)
 	}
 
-	ctx.AddFinalizer(func() error {
+	ctx.AddCleanUp(func() error {
 		err := client.BatchV1().Jobs(ctx.Namespace).Delete(context.TODO(), batch.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("failed to delete job: %v", err)
@@ -525,22 +517,4 @@ func run(client kubernetes.Interface, ctx *ScenarioContext, image string, name s
 	}
 
 	return nil
-}
-
-func CreateNamespace(client kubernetes.Interface, name string) error {
-	ns := &v1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-	}
-
-	_, err := client.CoreV1().Namespaces().Create(context.TODO(), ns, metav1.CreateOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to create namespace with name %v", name)
-	}
-	return nil
-}
-
-func DeleteNamespace(client kubernetes.Interface, name string) error {
-	return client.CoreV1().Namespaces().Delete(context.TODO(), name, metav1.DeleteOptions{})
 }